extract User into `lib/entities/user.dart`
