+++ lib/entities/user.dart
@@ -1,0 +1,18 @@
+import 'package:cloud_firestore/cloud_firestore.dart';
+
+class User {
+  final String id;
+  final String name;
+  final String phone;
+
+  User({required this.id, required this.name, required this.phone});
+
+  factory User.fromFirestore(DocumentSnapshot doc) {
+    Map<String, dynamic> data = doc.data() as Map<String, dynamic>;
+    return User(
+      id: doc.id,
+      name: data['name'] ?? '',
+      phone: data['phone'] ?? '',
+    );
+  }
+}
